requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-community>=0.2.0
huggingface-hub>=0.23.0
python-dotenv>=1.0.0
python-docx>=1.1.0
langchain-huggingface>=0.1.0
streamlit>=1.35.0